Internal: rename Touchable.css to TapArea.css
